backend/flask_app/db.py
import mysql.connector

def get_connection():
    return mysql.connector.connect(
        host="localhost",
        user="your_user",
        password="your_password",
        database="banking"
    )
